+++ lib/lite/uxid/record/hashid.rb
@@ -1,23 +1,23 @@
 # frozen_string_literal: true
 
-require "active_support" unless defined?(ActiveSupport)
-
 module Lite
   module Uxid
     module Record
       module Hashid
 
-        extend ActiveSupport::Concern
-
-        included do
-          after_commit :callback_generate_uxid!,
-                       if: proc { respond_to?(:uxid) && !uxid? },
-                       on: :create
+        def self.included(base)
+          base.extend ClassMethods
+          base.class_eval do
+            after_commit :callback_generate_uxid!,
+                         if: proc { respond_to?(:uxid) && !uxid? },
+                         on: :create
+          end
         end
 
-        class_methods do
+        module ClassMethods
+
           def find_by_uxid(uxid)
-            decoded_id = Lite::Uxid::Hashid.decode(uxid, prefix: new.uxid_prefix)
+            decoded_id = Lite::Uxid::Reversible::Hashid.decode(uxid, prefix: new.uxid_prefix)
             find_by(id: decoded_id)
           end
 
@@ -27,18 +27,19 @@ module Lite
 
             raise ActiveRecord::RecordNotFound
           end
+
         end
 
         def id_to_uxid
           return unless respond_to?(:uxid)
 
-          Lite::Uxid::Hashid.encode(id, prefix: uxid_prefix)
+          Lite::Uxid::Reversible::Hashid.encode(id, prefix: uxid_prefix)
         end
 
         def uxid_to_id
           return unless respond_to?(:uxid)
 
-          Lite::Uxid::Hashid.decode(uxid, prefix: uxid_prefix)
+          Lite::Uxid::Reversible::Hashid.decode(uxid, prefix: uxid_prefix)
         end
 
         def uxid_prefix