+++ lib/lite/uxid/record/ulid.rb
@@ -1,23 +1,21 @@
 # frozen_string_literal: true
 
-require "active_support" unless defined?(ActiveSupport)
-
 module Lite
   module Uxid
     module Record
       module Ulid
 
-        extend ActiveSupport::Concern
-
-        included do
-          before_create :callback_generate_uxid!,
-                        if: proc { respond_to?(:uxid) && !uxid? }
+        def self.included(base)
+          base.class_eval do
+            before_create :callback_generate_uxid!,
+                          if: proc { respond_to?(:uxid) && !uxid? }
+          end
         end
 
         private
 
         def callback_generate_uxid!
-          self.uxid = Lite::Uxid::Ulid.encode
+          self.uxid = Lite::Uxid::Irreversible::Ulid.encode
         end
 
       end