+++ spec/lite/uxid/irreversible/nanoid_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-RSpec.describe Lite::Uxid::Nanoid do
+RSpec.describe Lite::Uxid::Irreversible::Nanoid do
   let(:prefix) { nil }
   let(:encoder) { described_class.new(prefix: prefix) }
 
@@ -16,6 +16,7 @@ RSpec.describe Lite::Uxid::Nanoid do
       let(:prefix) { "test_" }
 
       it 'to be "test_1zWr1m0"' do
+        pp encoder.encode
         expect(described_class.encode(prefix: prefix).starts_with?(prefix)).to be(true)
         expect(encoder.encode.starts_with?(prefix)).to be(true)
       end