+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-09-23
+### Changed
+- Reorganize generators into better namespaces
+- Fixed issue where nanoid would return without the prefix
+
 ## [1.5.2] - 2024-09-20
 ### Changed
 - Fixed nanoid of differing length error